requirements.txt
requests==2.2.0
python-dateutil==2.2
pytz==2013.9